backend/requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
psycopg2-binary
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
celery
redis
requests
python-multipart
